inflearn/jpashop/src/main/java/io/ordi/jpabook/jpashop/MemberRepository.java
package io.ordi.jpabook.jpashop;

import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class MemberRepository {

    @PersistenceContext
    private EntityManager em;

    /**
     * Id 만 반환하는 이유
     * 커맨드와 쿼리를 분리해라 -> 사이드 이펙트 방지용
     */
    public Long save(Member member) {
        em.persist(member);
        return member.getId();
    }

    public Member find(Long id) {
        return em.find(Member.class, id);
    }
}
