inflearn/jpashop/src/test/java/io/ordi/jpabook/jpashop/MemberRepositoryTest.java
package io.ordi.jpabook.jpashop;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.jupiter.api.Assertions.*;

/**
 * https://effectivesquid.tistory.com/entry/Spring-Boot-starter-test-%EC%99%80-Junit5%EB%A5%BC-%EC%9D%B4%EC%9A%A9%ED%95%9C-%ED%85%8C%EC%8A%A4%ED%8A%B8
 */
@ExtendWith(SpringExtension.class)
@SpringBootTest(classes = JpashopApplication.class)
class MemberRepositoryTest {

    @Autowired
    MemberRepository memberRepository;

    @Test
    @Transactional // Spring 것 사용
    @Rollback(false) // 테스트 데이터 롤백 따로 하지 않음
    void save() {
        // given
        Member member = new Member();
        member.setUsername("memberA");

        // when
        Long savedId = memberRepository.save(member);
        Member findMember = memberRepository.find(savedId);

        // then
        Assertions.assertThat(findMember.getId())
                .isEqualTo(member.getId());
        Assertions.assertThat(findMember.getUsername())
                .isEqualTo(member.getUsername());
        Assertions.assertThat(findMember)
                .isEqualTo(member);
    }

}